add explicit link to Impressum
the Impressum should link to a wiki-page, that contains the actual
Impressum. There might be a problem with this, because everybody with
edit rights to the wiki can change the impressum, now. Therefore the
wiki needs a "lock"-feature someday. So a wiki-page might be locked to a
specific version or something like that. Anyway, this problem should be
handled elsewhere, not at this point.

This targets Issue MetaMeute/anduin#8