add some space between elements
This change aims for better separation of the content from navigational
elements.

The footer is not fixed at the bottom of the viewport any longer.
There’s no benefit in having it waste the space there. It’s now just at
the bottom of the page, so people will have to scroll down to see it.
This should be fine, though.